refactor: change structure for store
