+++ src/app/providers/storeProvider/config/StateSchema.ts
@@ -1,0 +1,4 @@
+export interface StateSchema {
+  // Here you will need to import schemas from slices
+  // In format user: UserSchema
+}
