--- src/app/providers/storeProvider/config/combineReducers/StateSchema.ts
@@ -1,3 +1,0 @@
-export interface StateSchema {
-
-}
