--- src/app/providers/storeProvider/config/combineReducers/index.ts
@@ -1,3 +1,0 @@
-export { rootReducer } from "./rootReducer";
-export { StateSchema } from "./StateSchema";
-
