--- src/app/providers/storeProvider/config/combineReducers/rootReducer.ts
@@ -1,1 +1,0 @@
-export const rootReducer = []
