+++ src/app/providers/storeProvider/index.ts
@@ -1,4 +1,3 @@
-import { StoreProvider } from "./ui/storeProvider";
-
-export default StoreProvider
-
+export { StoreProvider } from './ui/StoreProvider';
+export { StateSchema } from './config/StateSchema';
+export { RootState, AppDispatch } from './config/store';
