+++ src/app/providers/storeProvider/ui/storeProvider.tsx
@@ -1,11 +1,11 @@
-import { PropsWithChildren, ReactNode } from 'react';
-import { store } from '../config/store';
+import { ReactNode } from 'react';
 import { Provider } from 'react-redux';
+import { store } from '../config/store';
 
-export const StoreProvider = (children: PropsWithChildren) => {
-  return(
-    <Provider store={store}>
-      {children as ReactNode}
-    </Provider>
-  )
+interface StoreProviderProps {
+  children: ReactNode;
+}
+
+export const StoreProvider = ({ children }: StoreProviderProps) => {
+  return <Provider store={store}>{children}</Provider>;
 };
