+++ src/shared/lib/hooks/useAppDispatch/useAppDispatch.ts
@@ -1,4 +1,4 @@
-import { AppDispatch } from 'app/providers/storeProvider/config/store';
 import { useDispatch } from 'react-redux';
+import { AppDispatch } from 'app/Providers/StoreProvider';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
