+++ src/shared/lib/hooks/useAppSelect/useAppSelect.ts
@@ -1,4 +1,4 @@
-import { RootState } from 'app/providers/storeProvider/config/store';
 import { useSelector, TypedUseSelectorHook } from 'react-redux';
+import { RootState } from 'app/Providers/StoreProvider';
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
